+++ src/com/winterwell/juice/Juice.java
@@ -25,25 +25,6 @@ public class Juice {
 	private static final String LOGTAG = "juice";
 	List<AJuicer> juicers;
 	
-	/**
-	 * Just a test sketch of how to use this.
-	 * @param args
-	 */
-	public static void main(String[] args) {		
-		System.out.println("Juice version "+JuiceConfig.version); // TODO use config and AMain
-		Juice juice = new Juice();
-		if (args==null || args.length==0) {
-			args = new String[]{"http://www.guardian.co.uk/technology/2012/feb/29/raspberry-pi-computer-sale-british"};
-		}
-		
-		for (String url : args) {
-			String html = new FakeBrowser().getPage(url);
-			JuiceMe juiced = juice.juice(url, html);
-			Item item = juiced.getExtractedItems().get(0);
-			System.out.println('"'+item.getTitle()+"\" by "+item.getAuthor()+" date:"+item.getPublishedTime());
-		}
-	}
-	
 	public Juice() {
 		// in rough order of reliability
 		this(new ArrayList<> (Arrays.asList(
