+++ src/com/winterwell/juice/JuiceConfig.java
@@ -1,12 +1,18 @@
 package com.winterwell.juice;
 
+import com.winterwell.web.app.BasicSiteConfig;
+
 /**
  * STATUS: not used TODO
  * @author daniel
  *
  */
-public class JuiceConfig {
+public class JuiceConfig extends BasicSiteConfig {
 
+	public JuiceConfig() {
+		port = 8343;
+	}
+	
 	public static String version = "0.2.0";
 	
 }
